+++ source/features/stake-pools/actions.ts
@@ -1,0 +1,5 @@
+import Action from '../../utils/Action';
+
+export class StakePoolsActions {
+  public handleAcceptUnmoderatedData: Action<any> = new Action();
+}