+++ stories/header.stories.tsx
@@ -1,4 +1,5 @@
 import { storiesOf } from '@storybook/react';
+import { noop } from 'lodash';
 import React from 'react';
 import { BrandType } from '../source/common/constants';
 import Header from '../source/features/widgets/header/components/Header';
@@ -6,4 +7,10 @@ import { PaddingDecorator } from './support/PaddingDecorator';
 
 storiesOf('Header', module)
   .addDecorator(story => <PaddingDecorator>{story()}</PaddingDecorator>)
-  .add('Header', () => <Header withSearch brandType={BrandType.ENLARGED} />);
+  .add('Header', () => (
+    <Header
+      withSearch
+      brandType={BrandType.ENLARGED}
+      searchProps={{ onSearch: noop }}
+    />
+  ));